Abstract mock express app to its own module
With every router suite needing the mock app, abstract creation and
setup to a separate module that can be required as needed.

From research, middleware is located in the internal property
`app._router.stack` as an array, so `pop()`ing the stack in
`suiteTeardown()` with the understanding this will remove
the currently tested middleware/route from the app.