Testing (#10)
* Separate server.js into areas of responsibility

Throughout the creation of the api, server.js has had all the setup and
middleware logic continuously added to it. Main areas of responsibility
identified and extracted are:

1. Express application
2. Mongo/mongoose initialization and connection
3. Session storage middleware

* Add jest to project

* Add mocha and sinon

* Change mongoose model exports to set to
existing model if available before reinitializing

* Refactor to await database connection

* Add a few tests cover application initialization

* Add unit tests for authentication middleware

* Fix linting

* Save linter changes - single to double quotes

* Save linter changes

* Save linter changes and remove nonexistent field

* Configure prettier to work with eslint

* Formatting

- conform to prettier
- use single quotes again

* Update mongoose test to confirm a connected status

* Fix typo

* Format per prettier standards

* Add "inahand" to dictionary

* Assert mongoose test returns connection object

Prior test asserted readyState value. This must have been a side effect
of an existing connection, as readyState would not be set in the test,
as the connection is mocked.

* Refactor express app creation logic

Requiring app.js caused side effect of connecting to mongoose, resulting
in test results changing based on if the actual connection had occurred.
Refactor app.js to return the init function, and in server use the
returned express app when the promise resolves.

* Remove unnecessary server.test.js

The test is confirming require() works correctly, leaving that to node
to handle.

* Refactor scripts for prettier styling

* Refactor auth route

- move req/res logic into route
- unit test authentication logic

* Add a unit test for create user router endpoint

- add supertest to dev dependencies
- create express application with user router only; remember to
  use bodyparser middleware

* Refactor and fix wording

Move router use statement outside of setup as needed done once. Change
some wording to eliminate spellcheck hassle.

* Test /auth/login endpoint

- refactored session saving to its own service and converted return to
Promise
- had to mock session middleware to set session object on res

* Rename and refactor test

* Test logout endpoint

- move destroy session logic to its own service to return a Promise
- refactor endpoint to use async/await

* Test the auth check endpoint

Move session validation check to session service to encapulate logic

* Abstract mock express app to its own module

With every router suite needing the mock app, abstract creation and
setup to a separate module that can be required as needed.

From research, middleware is located in the internal property
`app._router.stack` as an array, so `pop()`ing the stack in
`suiteTeardown()` with the understanding this will remove
the currently tested middleware/route from the app.

* Move test-specific middleware out of the mock app
module and handle in `suiteSetup()` where needed

* Remove unused require

* Add test for find all accounts endpoint

Refactor the service to move request and response
logic to the router.

* Refactor auth middleware and user service
to use non-populated Identity document

* Refactor and test get all accounts route

* Refactor and test add account route

* Refactor account routing logic

- Move business logic from route to service
- Write tests

Co-authored-by: j-blake <j-blake@users.noreply.github.com>