Handle world currencies with ISO-4217 spec
Finally discovered ISO-4217, which is the governing body for world
currencies, although crypto is not currently part of it. Store the full
dataset as json and access through a promise to make any future api
calls easier to implement with Promise return signature already in
place. Use data to convert currencies to the most minor unit in order to
store transaction amounts as inteters instead of floats. Write tests.