[IN-12] Implement transactions (#26)
* IN-12 : Convert transaction route to typescript

* Convert transaction logic to typescript

* Update mongoose

* Create transaction repository

* Update payer/payee naming

Change to paying/receiving for easier reading and general clarity.

* Add before each hooks to reset mongoose models

* Handle world currencies with ISO-4217 spec

Finally discovered ISO-4217, which is the governing body for world
currencies, although crypto is not currently part of it. Store the full
dataset as json and access through a promise to make any future api
calls easier to implement with Promise return signature already in
place. Use data to convert currencies to the most minor unit in order to
store transaction amounts as inteters instead of floats. Write tests.

* Fix mongoose models typing, naming, transforming

* Create findOne repo function

* Refactor transaction service class to functions

Refactor find all, find one, and create methods to exported functions.

* Rename function name to be plural

* Implement transaction patch and delete endpoints

Co-authored-by: Blake McAskill <j-blake@users.noreply.github.com>