Create Docker configs
Attempt to containerize the application. Create Dockerfile.dev to use
nodemon to watch for changes and hot reload. Create Dockerfile for
production use; create multi-stage build process to compile typescript
then copy over just the necessary dist directory. Remove routes to
swagger as that is unused and way out of date. Keep the yml for now
as the documenation could be useful later.