Update all the things
Run `ncu -u` and get to fixing stuff. Mongoose options no longer needed
or supported, so remove.

Actually a nice opportunity to
review documentation. Mongoose is always a pain, but types seem a little
easier to handle. Reading through docs shows basically all the models
are created in a way not recommended. Next task will be to fix the
mongoose models, which in turn will hopefully make handing them off to
typed POJOs be saner.